First test passed!
This iteration of the interactive code contains a visualization of the
susceptibility distribution after manual segmentation of lungs,
vertebrae, ribs, esophagus, trachea, adrenal gland, thyroid gland, air
and fat.

This marks the development of the phantom as it works but it still needs
revision of all the methods as well as defining that all the organs have
susceptibility value of water and not 0.

It also needs to implement functions to save the np.arrays as niftis!